week5/requirements.txt
flask
redis
pytest
psycopg2-binary
